refactor: account fetches
