+++ webapp/src/pages/account/administrator/audits.tsx
@@ -1,26 +1,16 @@
 import { useEffect, useState } from "react";
-import { Table, Timeline, Button, Space, Input } from "antd";
+import { Table, Input } from "antd";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faArchive,
-  faBars,
-  faSearch,
-  faSignOutAlt,
-  faUser,
-} from "@fortawesome/free-solid-svg-icons";
+import { faSearch } from "@fortawesome/free-solid-svg-icons";
 import type { ChangeEvent } from "react";
 
-import { getAudits, getUserdata } from "#api/index";
-import { Header } from "#components/header";
-import { VideoList } from "#components/video-list";
-import Paginatione from "#components/pagination";
+import { getAudits } from "#api/index";
 
 function AuditsPage() {
   const [pageData, setPageData] = useState<{
     Events: Record<string, unknown>[];
     Length: number;
   } | null>(null);
-  const [userData, setUserData] = useState(null);
   const [userID, setUserID] = useState(-1);
   const [currPage, setPage] = useState(1);
 
@@ -58,9 +48,6 @@ function AuditsPage() {
     let fetchData = async () => {
       let data = await getAudits(userID, currPage);
       if (!ignore) setPageData(data);
-
-      let userData = await getUserdata();
-      if (!ignore) setUserData(userData);
     };
 
     fetchData();
@@ -73,8 +60,6 @@ function AuditsPage() {
 
   return (
     <>
-      <Header userData={userData} />
-
       <Input
         name="search"
         size="large"
